Implement Amazon EC2 blueprint for Drydock
Summary:
Ref T2015.  Depends on D10394.

This implements an Amazon EC2 blueprint that allocates and leases hosts from EC2. It starts EC2 instances according to the parameters specified on the blueprint, using the public key of a Passphrase key as the key pair for AWS.  For Linux, EC2 automatically configures the appropriate public key, and for Windows, this uses the userdata to automatically install and set up the Cygwin SSH server.  This means users can use the default Amazon instances with Drydock / Harbormaster and can connect out-of-the-box without any additional configuration.

It handles allocation of Elastic IPs, which provides a public IP address for connecting to the allocated machine when Phabricator is not hosted in the same VPC (or on Amazon).

This is built to allocate into a VPC; it doesn't support allocating a classic EC2 instance (and mainly because AWS is trying to deprecate that model anyway).

Test Plan: Tested with both Windows and Linux AMIs in the us-west-2 region.

Reviewers: epriestley, #blessed_reviewers

Subscribers: epriestley, Korvin

Maniphest Tasks: T2015

Differential Revision: https://secure.phabricator.com/D10395