Prevent allocation race by allocating lease before leaving lock
Summary: Resolves T6074.  This attempts to prevent an allocation race from occurring by saving the lease against the resource before we exit the lock.

Test Plan: Tested that allocations still work inside Harbormaster locally, but can't be entirely sure that the issue is resolved until we leave this fix in production for a little while.

Reviewers: #blessed_reviewers, epriestley

Subscribers: epriestley, Korvin

Maniphest Tasks: T6074

Differential Revision: https://secure.phabricator.com/D10470