Fix custom blueprint attributes to work when no attributes are specified
Summary: Ref T2015.  This fixes custom attribute parsing so that it works when the attribute text is empty (previously it would never match if there was empty text for the requirements and provided attributes).

Test Plan: Tested allocation of a resource when the blueprint attributes and lease host requirements set to empty text.

Reviewers: #blessed_reviewers, epriestley

Subscribers: epriestley, Korvin

Maniphest Tasks: T2015

Differential Revision: https://secure.phabricator.com/D10557