[drydock/core] Add blueprint PHID to Drydock logs
Summary: This adds a blueprint PHID to Drydock logs, so that logs can be associated against Drydock logs.  This is useful for tracking down issues when the Drydock allocator is choosing what blueprint to use, and can be used to log information about why blueprints were included or excluded from the available blueprints.

Test Plan: Brought this forward from D10506.  Rebased on upstream and allocated a lease against a preallocated resource blueprint.

Reviewers: epriestley, #blessed_reviewers

Subscribers: joshuaspence, Korvin, epriestley

Differential Revision: https://secure.phabricator.com/D10892