[drydock/core] Allow execution timeout to be set when running commands
Summary: Ref T2015.  This allows a maximum execution timeout to be set when running commands through Drydock.  This change is pulled out of D10479, and this functionality is used when blueprints want to query the readiness of a host resource they've created, but don't want to stall forever if the remote host is not responsive.

Test Plan: Pulled out of D10479, which has been in production for 6+ months.

Reviewers: #blessed_reviewers, epriestley

Subscribers: joshuaspence, Korvin

Maniphest Tasks: T2015

Differential Revision: https://secure.phabricator.com/D13985