[drydock/core] Allow leases to request a specific resource
Summary: Ref T2015. This allows leases to request a specific resource when acquiring a lease.  This is required for working copy caches where they must acquire a lease on the same machine as the acquirer (or otherwise fail).

Test Plan: Tested as part of the working copy implementation.

Reviewers: #blessed_reviewers, epriestley

Subscribers: epriestley, Korvin

Maniphest Tasks: T2015

Differential Revision: https://secure.phabricator.com/D13504