[drydock/working-copies/v0] Implement the working copy blueprint
Summary:
Ref T2015.  This implements a working copy blueprint which leases a working copy of a repository, commit or Differential revision (using the staging repository).

This only supports Git repositories at the moment, as I only use Git repositories and don't have the capacity to test or implement support for Mercurial or Subversion.

This implementation also attempts to acquire a working copy cache before cloning the repository; if acquisition of a cache succeeds then it uses the path provided by the cache, otherwise it clones directly from the original URL.  It is expected that v0 will always fail cache acquisition since the working copy cache implementation will be in v1.

Test Plan: Tested on a development machine.  Created a working copy blueprint and leased against it using the Harbormaster build steps in the next revision.

Reviewers: epriestley, #blessed_reviewers

Subscribers: epriestley, Korvin

Maniphest Tasks: T2015

Differential Revision: https://secure.phabricator.com/D13505