[drydock/working-copies/v1-caching] Implement working copy cache blueprint
Summary: Ref T2015.  This implements the working copy cache blueprint.  This blueprint attempts to acquire a lease on the same host resource as the acquirer, and keeps a cached version of the repository around for use.  If the host lease or resource is released or closed, this blueprint automatically flags it's resource as destroyed, and the acquirer (the working copy blueprint) will fall back to a direct clone.

Test Plan: Tested using a Harbormaster build plan (the same one from D13506).

Reviewers: epriestley, #blessed_reviewers

Subscribers: epriestley, Korvin

Maniphest Tasks: T2015

Differential Revision: https://secure.phabricator.com/D13507