[drydock/core] Support transient Drydock leases
Summary:
This adds support for transient Drydock leases.  Transient Drydock leases are leases which will not hold the Drydock resource open when Drydock is evaluating whether or not it can close the resource.

This is in addition to the working copy cache implementation, because it's important that the working copy cache lease does not keep a host resource open if it otherwise would be closed.

Test Plan: Tested by adding logging into the preallocated host blueprint and verified that the lease count is 0 if there are only transient leases left on the host resource.

Reviewers: epriestley, dfglhdfgdfgkjh, #blessed_reviewers

Subscribers: epriestley, Korvin

Differential Revision: https://secure.phabricator.com/D13510