[drydock/core] Allow transient leases to be leased even if the resource is at it's maximum lease limit
Summary:
Ref T2015.  This allows transient leases to be acquired on a resource where the blueprint implements min/max/expiry logic, as transient leases will not keep the resource open any way.

This corrects the scenario where a resource has a maximum of 1 lease, and a working copy is attempted to be obtained against that resource.  If the working copy cache blueprint is enabled, then it will always fail without this patch because of the maximum of 1 lease.  It is undesirable to set the maximum to 2, because we don't want to permit more than one build to run on the same host resource.

Test Plan: N/A

Reviewers: epriestley, #blessed_reviewers

Subscribers: epriestley, Korvin

Maniphest Tasks: T2015

Differential Revision: https://secure.phabricator.com/D13521