[harbormaster/core] Allow build variables to be optional by prefixing them with '?'
Summary: Ref T1049.  When a build variable is prefixed with a question mark, it does not have to be present when the variables are merged in.  If the build variable is missing or an empty string, then it is omitted from the merged result entirely (it does not even cause quotes to be added).  This is important when you an optional parameter you want to pass to a command, but when the variable is empty you don't want the command to recognise an argument at all.

Test Plan: Disabled storage for unit tests and ran `arc unit`.  I can't run unit tests locally because setting up the unit test databases on my computer takes like 20 minutes and I have no idea why.

Reviewers: epriestley, #blessed_reviewers

Subscribers: joshuaspence, epriestley, Korvin

Maniphest Tasks: T1049

Differential Revision: https://secure.phabricator.com/D13651