[drydock/core] Only show active resources and leases in controllers
Summary: Ref T2015.  These controllers get quite impossible to view when there are thousands of closed leases or resources.  Instead, only show the active leases and resources so it's easier for users to evaluate whether or not it's safe to manually close a resource.

Test Plan: Viewed it.

Reviewers: #blessed_reviewers, epriestley

Subscribers: epriestley, Korvin

Maniphest Tasks: T2015

Differential Revision: https://secure.phabricator.com/D13686