[drydock/working-copies] Set owner for working copies and show handles on attributes
Summary: Ref T2015.  This makes the working copy blueprints set the owner of leases when they request leases.  It also adds some PHID detection to the attribute view of leases and resources, to show handle links when the attribute is likely to contain a PHID.

Test Plan: Viewed pages.

Reviewers: #blessed_reviewers, epriestley

Subscribers: epriestley, Korvin

Maniphest Tasks: T2015

Differential Revision: https://secure.phabricator.com/D13689