[harbormaster/abort-builds/run-build-plan] Ensure 'Run Build Plan' fails if a sub-build moves into the aborted status
Summary:
This is split out from D11870 so that D11870 applies cleanly on origin/master.  This is the change required to D9807 when both D9807 and D11870 have landed.

This ensures that D11870 correctly fails when a child build moves into the aborted status.

Test Plan: Split out from D11870.

Reviewers: #blessed_reviewers, epriestley

Subscribers: epriestley, Korvin

Differential Revision: https://secure.phabricator.com/D13509