[custom/styling] Allow customization of page view and application name on Login / Register
Summary:
This allows for customization of the PageView class and the displayed application name on the Login and Register pages, via two GET parameters, `__pageview` and `__appname`.

The use case for this is having users authenticating through a website that's built on the Phabricator infrastructure, where the page view used on the site doesn't match the default Phabricator style.  In order to ensure users don't get confused logging in (before we can explain why we use Phabricator for our backend), we set the __pageview and __appname variables when Nginx performs URL rewriting.

I realise this might be a pretty unusual use case, but I do think this is something that will eventually need to be tackled in the very long term when people might be using the Phabricator infrastructure for building web applications.  I don't know whether allowing the user to set `__pageview` is a security issue; you can only change it to page view classes that are already defined.

`__appname` is so that the pages can show "MyApp Registration" instead of "Phabricator Registration".

Test Plan: Viewed the `/auth/login` and `/auth/register` pages with `__pageview` and `__appname` set.

Reviewers: #blessed_reviewers, epriestley

Subscribers: epriestley, Korvin

Differential Revision: https://secure.phabricator.com/D8642