[custom/harbormaster/phragment] Allow users to choose what types of buildables they want to publish fragments for
Summary:
Supersedes D7750.

Essentially the same thing but on the new field infrastructure.  I did want to make this work for all steps while I was at it, but since the fields are now drawn from the step configuration, I'm not sure what the right way to go about this (should I modify the steps to include a call to the parent for "global fields"?  Or should I modify the build step edit to forcefully include the field there?)

Test Plan: Tested against combinations of commits and revisions

Reviewers: epriestley, #blessed_reviewers

Subscribers: epriestley, Korvin

Maniphest Tasks: T4884

Differential Revision: https://secure.phabricator.com/D8641