[custom/s3-direct] Allow Phabricator to redirect to Amazon S3 for file downloads
Summary: This allows users to configure Phabricator to redirect to Amazon S3 for file downloads.  This allows bandwidth costs to be mitigated to Amazon S3 and means that Phabricator doesn't need to download data from S3 to serve it back to the user.

Test Plan: Tested on https://code.redpointsoftware.com.au/.  Applied the patch, migrated all the file data, and played around with the configuration options (and tested both downloading files and requesting a Phragment list via getstate).

Reviewers: #blessed_reviewers, epriestley

Subscribers: joshuaspence, hach-que, epriestley, Korvin

Maniphest Tasks: T5479, T4884

Differential Revision: https://secure.phabricator.com/D8655