[custom/phragment/bugfix] Create preceding paths for fragments
Summary: This fixes an issue where you are publishing a file in Harbormaster to a path like "path/to/new/fragment.txt", and the preceding directories don't yet exist.  This ensures that whenever a fragment is created, the preceding directories are also created.

Test Plan: Tested on my server and it worked.

Reviewers: epriestley, #blessed_reviewers

Subscribers: epriestley, Korvin, hach-que

Maniphest Tasks: T5479, T4884

Differential Revision: https://secure.phabricator.com/D8671