[custom/phragment/file] Create file controller in Phragment
Summary:
This provides a controller similar to the ZIP controller, except that it's meant for us on fragments that you want to download as a file.  This controller provides a consistent URL (that can optionally contain a snapshot) which will be correct when and if a fragment gets updated, unlike the raw file link.

It also adds "Download Snapshot" on the snapshot controller, which links to the file controller with the snapshot name included.

Test Plan: Downloaded file-based fragments from the new controller.

Reviewers: epriestley, #blessed_reviewers

Subscribers: epriestley, Korvin, hach-que

Maniphest Tasks: T5479, T4884

Differential Revision: https://secure.phabricator.com/D8672